src/utils/requirements.txt
loguru>=0.6.0
robotframework-sshlibrary==3.8.0
